Minesweeper Basic Update
This update will generate any size grid given a coordinates in the Creator gameObject and will spit out any size grid with the valid amount of mines for the minesweeper grid according to this Reddit post.  (Press O to generate, if O is pressed after a grid is made, then the board will be re-generated and the old board will be deleted).

https://www.reddit.com/r/Minesweeper/comments/mdbequ/what_is_a_good_mine_ratio/