Start of Number Algo
As the name suggests, this update is the start to a non-functioning algorithm to count the number of mines surrounding safe spaces. In order to start it, I ended up making the random mine generator always produce a safe spot, but that is easily fixed, but need to figure out how to access the text component and edit it to develop the algorithm to detect the mines.