Major Update, fixed major bug, added assets and some text
I managed to fix the issue with the TMP text, looks like the issue was me not using Get Compoenent in Child, then added the logic for the mine and number, added some backround sprites, and added some buttons for quitting the application, resetting the level, and getting back and forth from the menu to the field and visa versa. However, currently only have one Minesweeper mode, need to fix this in the future and resize (and maybe replace mines and other things with better assets.